requirements.txt
# ML & Data
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.3
joblib==1.4.2
matplotlib==3.9.3
seaborn==0.13.2

# Backend
django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.3.1

# Production
gunicorn==21.2.0
whitenoise==6.7.0

# MLOps
mlflow==2.19.0
optuna==4.1.0
scipy==1.13.1
